Improve move generation efficiency.

+++ src/uo_util.c
@@ -44,7 +44,7 @@ void uo_rand_init(uint64_t seed)
 #if !defined(uo_popcount)
 #include <limits.h>
 
-int uo_popcount(uint64_t u64)
+uint8_t uo_popcount(uint64_t u64)
 {
   uint64_t v = u64;
   v = v - ((v >> 1) & (uint64_t)~(uint64_t)0 / 3);                                            // temp
